+++ hc/front/tests/test_filtering_rules.py
@@ -12,12 +12,16 @@ class FilteringRulesTestCase(BaseTestCase):
 
     def test_it_works(self):
         self.client.login(username="alice@example.org", password="password")
-        r = self.client.post(self.url, data={"subject": "SUCCESS", "methods": "POST"})
+        r = self.client.post(
+            self.url,
+            data={"subject": "SUCCESS", "methods": "POST", "manual_resume": "1"},
+        )
         self.assertRedirects(r, self.redirect_url)
 
         self.check.refresh_from_db()
         self.assertEqual(self.check.subject, "SUCCESS")
         self.assertEqual(self.check.methods, "POST")
+        self.assertTrue(self.check.manual_resume)
 
     def test_it_clears_method(self):
         self.check.method = "POST"
@@ -29,3 +33,25 @@ class FilteringRulesTestCase(BaseTestCase):
 
         self.check.refresh_from_db()
         self.assertEqual(self.check.methods, "")
+
+    def test_it_clears_subject(self):
+        self.check.subject = "SUCCESS"
+        self.check.save()
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.post(self.url, data={"methods": ""})
+        self.assertRedirects(r, self.redirect_url)
+
+        self.check.refresh_from_db()
+        self.assertEqual(self.check.subject, "")
+
+    def test_it_clears_manual_resume_flag(self):
+        self.check.manual_resume = True
+        self.check.save()
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.post(self.url, data={})
+        self.assertRedirects(r, self.redirect_url)
+
+        self.check.refresh_from_db()
+        self.assertFalse(self.check.manual_resume)